--- src/frontend/screens/LikedVideosScreen/LikedVideos.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const LikedVideosScreen = () => {
-  return <div style={{ color: "white" }}>This is LikedVideos Screen</div>;
-};
-
-export { LikedVideosScreen };
