+++ src/frontend/screens/SignUpScreen/SignUpScreen.js
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { AUTH_TOKEN, USERNAME } from "../../constants/authConstants";
 import { useAuth } from "../../contexts";
-import { signUpService } from "../../services/auth-services/signUpService";
+import { signUpService } from "../../services";
 import "./SignUpScreen.css";
 
 const SignUpScreen = () => {
